Added more testing of XMI serialization of instances.
