+++ examples/xmi_node/test3a.testmodel
@@ -1,0 +1,2 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<TestModel:Item xmi:version="2.0" xmlns:xmi="http://www.omg.org/XMI" xmlns:TestModel="http://www.example.org/TestModel" name="Item3a"/>
