+++ examples/xmi_node/test3b.testmodel
@@ -1,0 +1,2 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<TestModel:SpecialItem xmi:version="2.0" xmlns:xmi="http://www.omg.org/XMI" xmlns:TestModel="http://www.example.org/TestModel" name=" Special Item3b" value="42"/>
