+++ test/models/test5-instance3a.xmi
@@ -1,0 +1,2 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<test5:Item xmi:version="2.0" xmlns:xmi="http://www.omg.org/XMI" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:test5="http://www.example.org/test5" name="Item3a"/>
